Run network and reputation tests separate For solidity-coverage and normal truffle test runs, leave in the same same folder so they are both run
