+++ test/contracts-network/colony-recovery.js
@@ -2,9 +2,9 @@ import chai from "chai";
 import bnChai from "bn-chai";
 import { ethers } from "ethers";
 
-import { SPECIFICATION_HASH } from "../helpers/constants";
-import { web3GetStorageAt, checkErrorRevert } from "../helpers/test-helper";
-import { setupColonyNetwork, setupMetaColonyWithLockedCLNYToken, setupRandomColony } from "../helpers/test-data-generator";
+import { SPECIFICATION_HASH } from "../../helpers/constants";
+import { web3GetStorageAt, checkErrorRevert } from "../../helpers/test-helper";
+import { setupColonyNetwork, setupMetaColonyWithLockedCLNYToken, setupRandomColony } from "../../helpers/test-data-generator";
 
 const { expect } = chai;
 chai.use(bnChai(web3.utils.BN));
