+++ test/contracts-network/colony-token-integrations.js
