+++ test/reputation-system/client-core-functionality.js
