+++ test/reputation-system/happy-paths.js
